requirements.txt
pandas
numpy
apache-airflow
pyspark